Updated Schema and DomainValidators to check enums correctly. Fixed remaining test issues. All 549u, 317/319i tests pass.

+++ Solution/Fabric.Test.Integration/FabApiModify/XAttachEventor.cs
@@ -83,7 +83,7 @@ namespace Fabric.Test.Integration.FabApiModify {
 		/*--------------------------------------------------------------------------------------------*/
 		[TestCase(0)]
 		[TestCase(-1)]
-		public void ErrEventorDateTimeRange(byte pDateTime) {
+		public void ErrEventorDateTimeRange(long pDateTime) {
 			vDateTime = pDateTime;
 			TestUtil.CheckThrows<FabArgumentOutOfRangeFault>(true, TestGo);
 		}
